+++ test/mpfr.jl
@@ -1022,3 +1022,14 @@ end
         @test big(typeof(complex(x, x))) == typeof(big(complex(x, x)))
     end
 end
+
+@testset "precision base" begin
+    setprecision(53) do
+        @test precision(Float64, base=10) == precision(BigFloat, base=10) == 15
+    end
+    setprecision(100, base=10) do
+        @test precision(BigFloat, base=10) == 100
+        @test precision(BigFloat, base=100) == 50
+        @test precision(BigFloat) == precision(BigFloat, base=2) = 333
+    end
+end